server/requirements.txt
python-nmap==0.6.0
python-socketio==1.0
psycopg2==2.7.5
